vm run_script and size
