+++ src/c_aci_testing/bicep/containerplatVM.bicep
@@ -8,6 +8,7 @@ param vmImage string
 param managedIDName string
 param containerplatUrl string
 param vmCustomCommands array = []
+param vmSize string = 'Standard_DC4ads_cc_v5'
 
 var tokenUrl = 'http://169.254.169.254/metadata/identity/oauth2/token?api-version=2018-02-01&resource=https://storage.azure.com/&client_id=${managedIdentity.properties.clientId}'
 
@@ -166,7 +167,7 @@ resource virtualMachine 'Microsoft.Compute/virtualMachines@2022-03-01' = {
   }
   properties: {
     hardwareProfile: {
-      vmSize: 'Standard_DC4ads_cc_v5'
+      vmSize: vmSize
     }
     storageProfile: {
       osDisk: {
